A major update to HIP DataModel:
1. Entities in ODS.DT are SerialObjects to be embedded in other entities
2. Other Entities in ODS package are persistent and now used in message
transferring only
3. Package HIP.DataModel were added, entities in this package inherit
from those in ODS package and to be used in persistent operations
4. Should be noticed that entity references in package HIP.DataModel
should all based on those in this package, which need to be done by
overriding. I.e. ODS.Encounter has a property as Subject referencing
ODS.Patient. In HIP.DataModel.Encounter, by inherintence,  it also has a
property as Subject referencing ODS.Patient, which should be overrided
as referencing to HIP.DataModel.Patient.